requirements.txt.txt
discord.py==1.6.0
urllib3==1.26.3
asyncio==3.4.3
beautifulsoup4==4.9.3